Add file structure + todo for the development
Separate js script for debugging in dev tools.

Todo gives a high level overview of what needs to
be done, which functions need to be created and what needs to be
returned.

The HTML will stay empty as of now, as it's a pure console log game
in its beginning stages (basic js practice)

README.md sets focus and goals for the project.